requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
idna==3.6
multidict==6.0.5
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.2
pydantic_core==2.16.3
pydentic==0.0.1.dev3
PyJWT==2.8.0
python-dotenv==1.0.1
python-stdnum==1.19
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
StrEnum==0.4.15
twilio==8.13.0
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
